+++ pymlg/torch/se3.py
@@ -140,7 +140,7 @@ class SE3(MatrixLieGroupTorch):
         )  # this yields a (N, 3, 4) matrix that must now be blocked with a (1, 4) batched matrix
 
         # generating a (N, 1, 4) batched matrix to append
-        b1 = torch.tensor([0, 0, 0, 0]).reshape(1, 1, 4)
+        b1 = torch.tensor([0, 0, 0, 0], dtype = xi.dtype).reshape(1, 1, 4)
         block = b1.repeat(Xi.shape[0], 1, 1)
 
         return torch.cat((Xi, block), dim=1)
@@ -187,7 +187,7 @@ class SE3(MatrixLieGroupTorch):
     
     @staticmethod
     def odot(b : torch.Tensor):
-        X = torch.zeros(b.shape[0], 4, 6)
+        X = torch.zeros(b.shape[0], 4, 6, dtype=b.dtype)
         X[:, 0:3, 0:3] = SO3.odot(b[0:3])
         X[:, 0:3, 3:6] = b[:, 3] * batch_eye(b.shape[0], 3, 3, dtype=b.dtype)
         return X
@@ -204,7 +204,7 @@ class SE3(MatrixLieGroupTorch):
     
     @staticmethod
     def adjoint_algebra(Xi):
-        A = torch.zeros(Xi.shape[0], 6, 6)
+        A = torch.zeros(Xi.shape[0], 6, 6, dtype=Xi.dtype)
         A[:, 0:3, 0:3] = Xi[:, 0:3, 0:3]
         A[:, 3:6, 0:3] = SO3.wedge(Xi[:, 0:3, 3])
         A[:, 3:6, 3:6] = Xi[:, 0:3, 0:3]
